social_bots/telegram_bot/requirements.txt
python-telegram-bot
python-dotenv
uvicorn
fastapi